Method put interactivo hecho, solucionar selección individual

+++ src/app/componentes/educacion/educacion.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { PortafolioService } from '../../services/portafolio.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-educacion',
@@ -14,7 +15,6 @@ export class EducacionComponent implements OnInit {
   form: FormGroup;
 
 
-
   constructor(public datosPortafolio: PortafolioService, private formBuilder: FormBuilder) {
     this.form=this.formBuilder.group({
      detalles: [ '', [Validators.required, Validators.minLength(2)]],
@@ -24,16 +24,22 @@ export class EducacionComponent implements OnInit {
      titulo: ['', [Validators.required, Validators.minLength(2)]]
 
     })
+    for(let i = 0; i < 100 ; i++) {
+      this.modoEdicion = false ;
+    }
    }
+
 
   ngOnInit(): void {
     this.datosPortafolio.obtenerDatosEducacion().subscribe(data => {
       console.log("Datos Personales: " + JSON.stringify(data));
       this.miPortafolio=data;
-      console.log(data);
+      console.log("data: ",data);
+      console.log("miPortafolio", this.miPortafolio);
     });
-
   }
+
+
 
 
   onCrear(event: Event){
@@ -42,23 +48,20 @@ export class EducacionComponent implements OnInit {
   }
 
 
-  onEdit(i: any, event: Event ){
-    this.modoEdicion=true;
-    event.preventDefault;
-    this.datosPortafolio.putEducacion(this.form.value, this.form.value[i].id).subscribe(data => {
-      console.log("this.form.value: " , this.form.value);
-      console.log("AcercaDe method post Data", data);
-    });
+  onEdit(id: any, event: Event ){
+    this.modoEdicion= true;
+    console.log("this.form.value: " , this.form.value);
+    console.log("id: " , id);
   }
 
-  onSaveNew(event: Event ){
+  onSaveEdit( id: any, event: Event ){
     event.preventDefault;
-    this.datosPortafolio.postEducacion(this.form.value).subscribe(data => {
+    this.datosPortafolio.putEducacion(this.form.value, id).subscribe(data => {
       console.log("this.form.value: " , this.form.value);
-      console.log("AcercaDe method post Data", data);
-    this.modoNuevoRegistro=false;
-    //this.ruta.navigate(['/portafolio']);
+      console.log("id: " , id);
+      console.log("EDUCACIÓN method PUT Data Editada", data);
     });
+    this.modoEdicion = false;
   }
 
   onSaveNewNuevoRegistro(event: Event ){
@@ -67,19 +70,41 @@ export class EducacionComponent implements OnInit {
       console.log("this.form.value: " , this.form.value);
       console.log("AcercaDe method post Data", data);
     this.modoNuevoRegistro=false;
-    //this.ruta.navigate(['/portafolio']);
+    });
+
+    this.datosPortafolio.obtenerDatosEducacion().subscribe(data => {
+      this.miPortafolio=data;
     });
   }
 
   onDelete(  i: any, event: Event ){
-    this.modoEdicion=false;
+    this.modoEdicion = false;
     event.preventDefault;
-    this.datosPortafolio.deleteEducacion(this.miPortafolio[i].id).subscribe(data => {
-    console.log("Borrando registro", data);
-    });
+    Swal.fire({
+      title: '¿Desea Eliminar el item de Educación?',
+      text: "No podrá revertir los cambios.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Si, Eliminar.'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.datosPortafolio.deleteEducacion(this.miPortafolio[i].id).subscribe(data => {
+          console.log("Borrando registro", data);
 
+          this.datosPortafolio.obtenerDatosEducacion().subscribe(data => {
+            this.miPortafolio=data;
+          });
 
-
+          });
+        Swal.fire(
+          'ELIMINADO',
+          'Item Educación eliminado con éxito.',
+          'success'
+        )
+      }
+    })
   }
 
   onCancelNuevoRegistro(){
@@ -87,7 +112,7 @@ export class EducacionComponent implements OnInit {
   }
 
   onCancel(i: any, event: Event){
-    this.modoEdicion=false;
+    this.modoEdicion= false;
   }
