+++ src/app/componentes/proyectos/proyectos.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { PortafolioService } from '../../services/portafolio.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-proyectos',
@@ -38,23 +39,20 @@ export class ProyectosComponent implements OnInit {
   }
 
 
-  onEdit(i: any, event: Event ){
+  onEdit(id: any, event: Event ){
     this.modoEdicion=true;
-    event.preventDefault;
-    this.datosPortafolio.putProyecto(this.form.value, this.form.value[i].id).subscribe(data => {
-      console.log("this.form.value: " , this.form.value);
-      console.log("PROYECTO method PUT Data Editada", data);
-    });
+    console.log("this.form.value: " , this.form.value);
+    console.log("id: " , id);
   }
 
-  onSaveNew(event: Event ){
+  onSaveEdit( id: any, event: Event ){
     event.preventDefault;
-    this.datosPortafolio.postProyecto(this.form.value).subscribe(data => {
+    this.datosPortafolio.putProyecto(this.form.value, id).subscribe(data => {
       console.log("this.form.value: " , this.form.value);
-      console.log("PROYECTO method POST Data Enviada", data);
-    this.modoEdicion=false;
-    //this.ruta.navigate(['/portafolio']);
+      console.log("id: " , id);
+      console.log("PROYECTO method PUT Data Editada", data);
     });
+    this.modoEdicion=false;
   }
 
   onSaveNewNuevoRegistro(event: Event ){
@@ -63,8 +61,12 @@ export class ProyectosComponent implements OnInit {
       console.log("this.form.value: " , this.form.value);
       console.log("PROYECTO method POST Data Enviada", data);
     this.modoNuevoRegistro=false;
-    //this.ruta.navigate(['/portafolio']);
     });
+
+    this.datosPortafolio.obtenerDatosProyectos().subscribe(data => {
+      this.miPortafolio=data;
+    });
+
   }
 
 
@@ -81,9 +83,33 @@ export class ProyectosComponent implements OnInit {
   onDelete( i: any, event: Event ){
     this.modoEdicion=false;
     event.preventDefault;
-    this.datosPortafolio.deleteProyecto(this.miPortafolio[i].id).subscribe(data => {
-    console.log("Borrando registro", data);
-    });
+    Swal.fire({
+      title: '¿Desea Eliminar el item de Proyectos?',
+      text: "No podrá revertir los cambios.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'ELIMINAR'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.datosPortafolio.deleteProyecto(this.miPortafolio[i].id).subscribe(data => {
+          console.log("Borrando registro", data);
+
+          this.datosPortafolio.obtenerDatosProyectos().subscribe(data => {
+            this.miPortafolio=data;
+          });
+
+          });
+
+        Swal.fire(
+          'ELIMINADO',
+          'Item de Proyectos eliminado con éxito.',
+          'success'
+        )
+      }
+    })
+
   }
